Revert a 0.5.15lorg2-alt70 change, which is too rude: "Support large files"
It touches a lot of types in the code without good justification and
overlooks places where values of incompatible integer types are
assigned or compared.

This reverts commit 297a12d922fbb9645c9f45801f1c3ef5517bf168.

+++ apt/apt-pkg/acquire.cc
@@ -781,7 +781,7 @@ bool pkgAcquireStatus::Pulse(pkgAcquire *Owner)
    }
    
    // Compute the current completion
-   unsigned long long ResumeSize = 0;
+   unsigned long ResumeSize = 0;
    for (pkgAcquire::Worker *I = Owner->WorkersBegin(); I != 0;
 	I = Owner->WorkerStep(I))
       if (I->CurrentItem != 0 && I->CurrentItem->Owner->Complete == false)
@@ -868,7 +868,7 @@ void pkgAcquireStatus::Stop()
 // AcquireStatus::Fetched - Called when a byte set has been fetched	/*{{{*/
 // ---------------------------------------------------------------------
 /* This is used to get accurate final transfer rate reporting. */
-void pkgAcquireStatus::Fetched(unsigned long long Size,unsigned long long Resume)
+void pkgAcquireStatus::Fetched(unsigned long Size,unsigned long Resume)
 {   
    FetchedBytes += Size - Resume;
 }
